feat(cli)!: fold the dataset read surface into the dataset group
`dataset` absorbs `query dates` (now `dataset dates NAME`, resolving active
or not) and `report`'s `coverage`/`value-ranges`/`grid` (now `dataset dates
--gaps`, `dataset values`, and grid details folded into `dataset info`).
`query.py` and `report.py` are deleted and unregistered.
