+++ tests/cli/test_stats_cli.py
@@ -1,4 +1,4 @@
-"""CLI tests for the top-level ``stats`` command and the ``query dates`` command.
+"""CLI tests for the top-level ``stats`` command.
 
 ``stats`` needs the full set of prerequisites -- a stored AOI, generated
 terrain + land cover, a burned AOI raster, and an ingested COG -- so a fixture
@@ -280,19 +280,3 @@ def test_stats_unknown_dataset_is_clean_error(runner, cli_obj, populated_root):
     )
     assert result.exit_code != 0
     assert 'No such dataset' in result.output
-
-
-def test_query_dates_lists_ingested_dates(runner, cli_obj, populated_root):
-    result = runner.invoke(cli, ['query', 'dates', '-d', 'test'], obj=cli_obj)
-    assert result.exit_code == 0, result.output
-    assert '2018-04-27' in result.output
-
-
-def test_query_dates_filters_by_range(runner, cli_obj, populated_root):
-    result = runner.invoke(
-        cli,
-        ['query', 'dates', '-d', 'test', '--start', '20190101'],
-        obj=cli_obj,
-    )
-    assert result.exit_code == 0, result.output
-    assert '2018-04-27' not in result.output
